requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langgraph
openai
tavily-python
python-dotenv
pillow
requests 